Feat : refactor project structures
