+++ src/pages/Staking/utils/index.ts
@@ -1,0 +1,3 @@
+import {fetchManageModalPayload} from './fetchManageModalPayload';
+
+export {fetchManageModalPayload};
